refactor(core): split per-stat frame builders and facet grid margins
Extract count/bin/density and smooth/boxplot/summary into focused frame-stats
modules behind stable facades; split facet placement into margin/geometry pass
and tick placement. Tests pin density singleton-group warnings and free_x axes.
